Use `composer update` rather than `install` when lockfile isn't checked in. (#19173)
If the lock file isn't checked in to git, we should use
`composer update` instead of using whatever stale lockfile happens to be
in place.

This makes that happen for the Jetpack CLI (for composer and yarn) and
for the composer scripts in various packages. It doesn't handle the CLI
"generate" command skeleton, as that currently doesn't seem to have
logic determining whether composer.lock is checked in or not. Thus this
doesn't completely fix #18753.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/661184144